+++ tests/testthat/test-result.R
@@ -1,5 +1,4 @@
 test_that("result field", {
-
   field <- new_result_field()
 
   expect_s3_class(field, "result_field")
@@ -8,7 +7,6 @@ test_that("result field", {
 })
 
 test_that("result block", {
-
   block <- new_result_block()
 
   expect_s3_class(block, "result_block")
@@ -19,8 +17,7 @@ test_that("result block", {
   expect_s3_class(ui, "shiny.tag")
 })
 
-test_that("result server works", {
-
+test_that("result field server works", {
   set_workspace(
     stack1 = new_stack(new_dataset_block),
     stack2 = new_stack(new_dataset_block),
@@ -28,8 +25,25 @@ test_that("result server works", {
   )
 
   shiny::testServer(
-    generate_server(new_result_field()), {
-      expect_setequal(opts(), c("stack1", "stack2"))
+    generate_server(new_result_field("stack1")),
+    {
+      expect_setequal(workspace_stacks(), c("stack1", "stack2"))
     }
+  )
+})
+
+withr::local_package("shinytest2")
+
+test_that("result server works", {
+  skip_on_cran()
+
+  app <- AppDriver$new(
+    system.file("examples/result/app.R", package = "blockr"),
+    name = "result-app"
+  )
+
+  app$expect_values(
+    input = "select-stack",
+    export = "stacks"
   )
 })